requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12.0,<4.0
psycopg2-binary>=2.8,<3.0  # For PostgreSQL database
